refactoring removed unnecessary "docker pull"
