+++ .travis.yml
@@ -4,7 +4,7 @@ services:
   - docker
 
 before_install:
-  - docker pull elasticsearch:7.3.2
+  # - docker pull elasticsearch:7.3.2
   - docker run -d -p 9200:9200 -p 9300:9300 -e "discovery.type=single-node" docker.elastic.co/elasticsearch/elasticsearch:7.3.2
   - docker ps -a
