+++ example.py
@@ -1,4 +1,4 @@
-from py_image_dedup.config.deduplicator_config import DeduplicatorConfig
+from py_image_dedup.config import DeduplicatorConfig
 from py_image_dedup.library.deduplicator import ImageMatchDeduplicator
 
 config = DeduplicatorConfig()
