--- py_image_dedup/config/__init__.py
@@ -1,24 +1,0 @@
-NODE_MAIN = "py_image_dedup"
-
-NODE_DRY_RUN = "dry_run"
-
-NODE_ELASTICSEARCH = "elasticsearch"
-
-NODE_HOST = "host"
-NODE_MAX_DISTANCE = "max_distance"
-NODE_AUTO_CREATE_INDEX = "auto_create_index"
-
-NODE_ANALYSIS = "analysis"
-
-NODE_SOURCE_DIRECTORIES = "source_directories"
-NODE_RECURSIVE = "recursive"
-NODE_SEARCH_ACROSS_ROOT_DIRS = "across_dirs"
-NODE_FILE_EXTENSIONS = "file_extensions"
-NODE_USE_EXIF_DATA = "use_exif_data"
-NODE_THREADS = "threads"
-
-NODE_DEDUPLICATION = "deduplication"
-
-NODE_MAX_FILE_MODIFICATION_TIME_DIFF = "max_file_modification_time_diff"
-NODE_REMOVE_EMPTY_FOLDERS = "remove_empty_folders"
-NODE_DUPLICATES_TARGET_DIRECTORY = "duplicates_target_directory"
