+++ tests/__init__.py
@@ -1,6 +1,6 @@
 import unittest
 
-from py_image_dedup.config.deduplicator_config import DeduplicatorConfig
+from py_image_dedup.config import DeduplicatorConfig
 
 
 class TestBase(unittest.TestCase):
